Finished with the Header section now to work on the banner.
So what I did here was I add a class to the header h1 so that I could target it by itself. Afterwards it was setting up the color codes to match the project portfolio.

Next I add all the right margin, padding, and colors to the matching blocks and text.